Moved pending files from /wordpress to /project_types/wordpress
