+++ project_types/wordpress/Install-WpCli.sh
