+++ project_types/wordpress/basic_structure_starter.py
