+++ project_types/wordpress/default-localhost-site.json
