+++ project_types/wordpress/default-project.xml
