+++ project_types/wordpress/default-wordpress-project-structure.json
