+++ project_types/wordpress/default.gitignore
